requirements.txt
# Solar Sense — Streamlit app dependencies
streamlit>=1.28.0
plotly>=5.18.0
pyserial>=3.5
pandas>=2.0.0
numpy>=1.24.0
